Converter, クラス名をMZ1Z001からMZ1Z002へ変更した
