+++ SBasicConverter/source/MZ1Z002.cpp
@@ -4,11 +4,11 @@
 #include <vector>
 #include <algorithm>
 #include "Format.hpp"
-#include "MZ1Z001.hpp"
+#include "MZ1Z002.hpp"
 #include "BasicCode.h"
 #include "FileData.hpp"
 
-MZ1Z001::MZ1Z001(void)
+MZ1Z002::MZ1Z002(void)
 :space(4)
 ,variableList()
 ,dimVariableList()
@@ -41,12 +41,12 @@ MZ1Z001::MZ1Z001(void)
 {
 }
 
-MZ1Z001::~MZ1Z001(void)
+MZ1Z002::~MZ1Z002(void)
 {
 }
 
 // データをバッファに格納
-void MZ1Z001::Load(char* buffer)
+void MZ1Z002::Load(char* buffer)
 {
 	this->buffer = buffer;
 	this->linePointer = 128;
@@ -59,7 +59,7 @@ void MZ1Z001::Load(char* buffer)
 }
 
 // 1行分のデータを取得
-bool MZ1Z001::GetLine(std::vector<char>& receiveBuffer, int& number)
+bool MZ1Z002::GetLine(std::vector<char>& receiveBuffer, int& number)
 {
 	receiveBuffer.clear();
 	number = 0;
@@ -82,7 +82,7 @@ bool MZ1Z001::GetLine(std::vector<char>& receiveBuffer, int& number)
 }
 
 // コンバート
-bool MZ1Z001::ConvertFile(const std::string filepath, const std::string outputFilepath)
+bool MZ1Z002::ConvertFile(const std::string filepath, const std::string outputFilepath)
 {
 	this->space = 4;
 	this->variableList.clear();
@@ -305,7 +305,7 @@ bool MZ1Z001::ConvertFile(const std::string filepath, const std::string outputFi
 }
 
 // 後ろに変数が続く可能性のある命令か
-bool MZ1Z001::IsExistVariable(std::string command)
+bool MZ1Z002::IsExistVariable(std::string command)
 {
 	std::vector<std::string> commandTable =
 	{
@@ -336,7 +336,7 @@ bool MZ1Z001::IsExistVariable(std::string command)
 }
 
 // 修正前の変数名かどうか(文字列には$がついている)
-bool MZ1Z001::IsVariableNameChar(unsigned char byte, bool first)
+bool MZ1Z002::IsVariableNameChar(unsigned char byte, bool first)
 {
 	if(byte >= 'A' && byte <= 'Z')
 	{
@@ -357,7 +357,7 @@ bool MZ1Z001::IsVariableNameChar(unsigned char byte, bool first)
 }
 
 // 修正後の変数名かどうか(文字列には$のかわりにstrが付いている)
-bool MZ1Z001::IsFixedVariableNameChar(unsigned char byte, bool first)
+bool MZ1Z002::IsFixedVariableNameChar(unsigned char byte, bool first)
 {
 	if(byte >= 'A' && byte <= 'Z')
 	{
@@ -379,7 +379,7 @@ bool MZ1Z001::IsFixedVariableNameChar(unsigned char byte, bool first)
 
 // 変数を登録する、変数名を変更する
 // 登録出来たらtrueを返す
-bool MZ1Z001::SetVariableName(std::string& variableName, std::vector<char>& result, bool array, int number)
+bool MZ1Z002::SetVariableName(std::string& variableName, std::vector<char>& result, bool array, int number)
 {
 	// 空なら何もしない
 	if(variableName.empty() == true)
@@ -459,7 +459,7 @@ bool MZ1Z001::SetVariableName(std::string& variableName, std::vector<char>& resu
 	return true;
 }
 
-bool MZ1Z001::IsBracketCommand(std::string command)
+bool MZ1Z002::IsBracketCommand(std::string command)
 {
 	if(command.find("(") == std::string::npos)
 	{
@@ -469,7 +469,7 @@ bool MZ1Z001::IsBracketCommand(std::string command)
 }
 
 // (カッコの始まりを登録する
-void MZ1Z001::PushBrackets(void)
+void MZ1Z002::PushBrackets(void)
 {
 	this->bracketsCountList.push_back(0);
 	for(size_t i = 0; i < this->bracketsCountList.size(); ++ i)
@@ -479,7 +479,7 @@ void MZ1Z001::PushBrackets(void)
 }
 
 // (カッコを見つけたらインクリメントする
-void MZ1Z001::AddBrackets(void)
+void MZ1Z002::AddBrackets(void)
 {
 	for(size_t i = 0; i < this->bracketsCountList.size(); ++ i)
 	{
@@ -489,7 +489,7 @@ void MZ1Z001::AddBrackets(void)
 
 // )カッコを見つけたらデクリメントして0ならtrueを返す
 // trueの場合は登録したカッコ(に対応した)ガッコが見つかった
-bool MZ1Z001::PopBrackets(void)
+bool MZ1Z002::PopBrackets(void)
 {
 	for(size_t i = 0; i < this->bracketsCountList.size(); ++ i)
 	{
@@ -504,7 +504,7 @@ bool MZ1Z001::PopBrackets(void)
 }
 
 // []カッコ処理中か
-bool MZ1Z001::IsBrackets(void)
+bool MZ1Z002::IsBrackets(void)
 {
 	for(size_t i = 0; i < this->bracketsCountList.size(); ++ i)
 	{
@@ -518,7 +518,7 @@ bool MZ1Z001::IsBrackets(void)
 
 // 配列のカッコを()から[]に変換する
 // 文字列変数名を$マーク付きからstr付きに修正する
-std::vector<char> MZ1Z001::PreConvertLine(const std::vector<char>& buffer, int number)
+std::vector<char> MZ1Z002::PreConvertLine(const std::vector<char>& buffer, int number)
 {
 	std::vector<char> result;
 	std::string variable;
@@ -859,7 +859,7 @@ std::vector<char> MZ1Z001::PreConvertLine(const std::vector<char>& buffer, int n
 // LINE_END: 行が終わるまで
 // SQUARE_BRACKETS_END: ]が検出されるまで
 // DELIMITER_END: :か行の最後まで
-bool MZ1Z001::Convert(const std::vector<char>& buffer, int number, int conditions)
+bool MZ1Z002::Convert(const std::vector<char>& buffer, int number, int conditions)
 {
 	this->space = 4;
 	Lexical lexical = { "NOP", "" };
@@ -1112,7 +1112,7 @@ bool MZ1Z001::Convert(const std::vector<char>& buffer, int number, int condition
 	return true;
 }
 
-void MZ1Z001::ConvertSub(const std::vector<char>& buffer, Lexical& lexical, unsigned char byte, bool processedDelimiter, int number, int& subNumber)
+void MZ1Z002::ConvertSub(const std::vector<char>& buffer, Lexical& lexical, unsigned char byte, bool processedDelimiter, int number, int& subNumber)
 {
 	this->debugLine += Format("%s", code_80_xx[byte]);
 	++ this->convertIndex;
@@ -1126,7 +1126,7 @@ void MZ1Z001::ConvertSub(const std::vector<char>& buffer, Lexical& lexical, unsi
 	-- this->convertIndex;
 }
 
-void MZ1Z001::AnalyzeCommand(Lexical& lexical, int number, int& subNumber, bool delimiter)
+void MZ1Z002::AnalyzeCommand(Lexical& lexical, int number, int& subNumber, bool delimiter)
 {
 	if (lexical.command == "NOP" && lexical.option == "")
 	{
@@ -1434,7 +1434,7 @@ void MZ1Z001::AnalyzeCommand(Lexical& lexical, int number, int& subNumber, bool 
 #endif
 }
 
-void MZ1Z001::Delimiter(int number, int& subNumber)
+void MZ1Z002::Delimiter(int number, int& subNumber)
 {
 	if(this->closeBracesFlag == true)
 	{
@@ -1469,7 +1469,7 @@ void MZ1Z001::Delimiter(int number, int& subNumber)
 	}
 }
 
-void MZ1Z001::AddColor(void)
+void MZ1Z002::AddColor(void)
 {
 	if(this->color.empty() == false)
 	{
@@ -1482,7 +1482,7 @@ void MZ1Z001::AddColor(void)
 	this->addColorFlag = false;
 }
 
-std::string MZ1Z001::Nop(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Nop(const Lexical& lexical, bool delimiter)
 {
 	std::string text = lexical.option;
 	if(text.empty() == true)
@@ -1493,14 +1493,14 @@ std::string MZ1Z001::Nop(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Rem(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Rem(const Lexical& lexical, bool delimiter)
 {
 	std::string comment = Encode(lexical.option);
 	std::string result = "/* " + comment + " */";
 	return result;
 }
 
-std::string MZ1Z001::Data(const Lexical& lexical, bool delimiter, int number)
+std::string MZ1Z002::Data(const Lexical& lexical, bool delimiter, int number)
 {
 	std::string data = ParseData(lexical.option);
 	std::string result = "SetData(D" + Number(number, 1) + ", " + data + ")";
@@ -1508,7 +1508,7 @@ std::string MZ1Z001::Data(const Lexical& lexical, bool delimiter, int number)
 	return "";
 }
 
-std::string MZ1Z001::DefKey(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::DefKey(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "DefKey(";
 	std::string option = Trim(lexical.option);
@@ -1543,7 +1543,7 @@ std::string MZ1Z001::DefKey(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Read(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Read(const Lexical& lexical, bool delimiter)
 {
 	// Readする変数リスト作成
 	std::vector<std::string> variableList = VariableList(lexical.option);
@@ -1568,7 +1568,7 @@ std::string MZ1Z001::Read(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-void MZ1Z001::Run(const Lexical& lexical, bool delimiter, int number, int& subNumber)
+void MZ1Z002::Run(const Lexical& lexical, bool delimiter, int number, int& subNumber)
 {
 	this->program[Number(number, subNumber)] = "Clr();";
 	++ subNumber;
@@ -1577,7 +1577,7 @@ void MZ1Z001::Run(const Lexical& lexical, bool delimiter, int number, int& subNu
 	this->result = "";
 }
 
-std::string MZ1Z001::Print(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Print(const Lexical& lexical, bool delimiter)
 {
 	this->printFlag = true;
 	this->printSemicolonFlag = false;
@@ -1603,7 +1603,7 @@ std::string MZ1Z001::Print(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-void MZ1Z001::EndPrint(int number, int& subNumber)
+void MZ1Z002::EndPrint(int number, int& subNumber)
 {
 	// Print文中の最後の;は;を消してフラグを立てる
 	if((this->result.empty() == false) && (this->result[this->result.size() - 1] == ';'))
@@ -1627,7 +1627,7 @@ void MZ1Z001::EndPrint(int number, int& subNumber)
 	this->printSemicolonFlag = false;
 }
 
-std::string MZ1Z001::For(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::For(const Lexical& lexical, bool delimiter)
 {
 	++ this->forStackCount;
 	this->forPhase = 1;
@@ -1646,7 +1646,7 @@ std::string MZ1Z001::For(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::To(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::To(const Lexical& lexical, bool delimiter)
 {
 	this->forPhase = 2;
 	std::string result = ", ";
@@ -1654,7 +1654,7 @@ std::string MZ1Z001::To(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Step(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Step(const Lexical& lexical, bool delimiter)
 {
 	this->forPhase = 3;
 	std::string result = ", ";
@@ -1662,7 +1662,7 @@ std::string MZ1Z001::Step(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-void MZ1Z001::EndFor(int number, int& subNumber)
+void MZ1Z002::EndFor(int number, int& subNumber)
 {
 	if(this->forPhase == 1)
 	{
@@ -1685,7 +1685,7 @@ void MZ1Z001::EndFor(int number, int& subNumber)
 	this->forPhase = 0;
 }
 
-void MZ1Z001::Next(const Lexical& lexical, bool delimiter, int number, int& subNumber)
+void MZ1Z002::Next(const Lexical& lexical, bool delimiter, int number, int& subNumber)
 {
 	std::vector<std::string> variableList = VariableList(lexical.option);
 	if(variableList.size() == 0)
@@ -1705,7 +1705,7 @@ void MZ1Z001::Next(const Lexical& lexical, bool delimiter, int number, int& subN
 	}
 }
 
-std::string MZ1Z001::If(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::If(const Lexical& lexical, bool delimiter)
 {
 	++ this->ifStackCount;
 	std::string result = "If(";
@@ -1715,7 +1715,7 @@ std::string MZ1Z001::If(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-void MZ1Z001::EndIf(int number, int& subNumber)
+void MZ1Z002::EndIf(int number, int& subNumber)
 {
 	this->result += ", {L_NEXT_COMMAND}, {L_NEXT_LINE});";
 	this->program[Number(number, subNumber)] = this->result;
@@ -1724,7 +1724,7 @@ void MZ1Z001::EndIf(int number, int& subNumber)
 	-- this->ifStackCount;
 }
 
-std::string MZ1Z001::Then(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Then(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "";
 	std::string option = Trim(lexical.option);
@@ -1743,7 +1743,7 @@ std::string MZ1Z001::Then(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Goto(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Goto(const Lexical& lexical, bool delimiter)
 {
 	std::string result;
 	if(this->onFlag == true)
@@ -1760,7 +1760,7 @@ std::string MZ1Z001::Goto(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Gosub(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Gosub(const Lexical& lexical, bool delimiter)
 {
 	std::string result;
 	if(this->onFlag == true)
@@ -1777,39 +1777,39 @@ std::string MZ1Z001::Gosub(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Return(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Return(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Return()";
 	return result;
 }
 
-std::string MZ1Z001::Stop(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Stop(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Stop()";
 	return result;
 }
 
-std::string MZ1Z001::End(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::End(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "End()";
 	return result;
 }
 
-std::string MZ1Z001::On(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::On(const Lexical& lexical, bool delimiter)
 {
 	this->onFlag = true;
 	this->onVariable = FixOptionNumber(lexical.option);
 	return "";
 }
 
-std::string MZ1Z001::Poke(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Poke(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Poke(" + FixOptionNumber(lexical.option);
 	this->closeBracketFlag = true;
 	return result;
 }
 
-void MZ1Z001::Dim(const Lexical& lexical, bool delimiter, int number, int& subNumber)
+void MZ1Z002::Dim(const Lexical& lexical, bool delimiter, int number, int& subNumber)
 {
 	std::vector<std::string> variableList = DimVariableList(lexical.option);
 	for(size_t i = 0; i < variableList.size(); ++ i)
@@ -1867,7 +1867,7 @@ void MZ1Z001::Dim(const Lexical& lexical, bool delimiter, int number, int& subNu
 	this->result = "";
 }
 
-std::string MZ1Z001::DefFn(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::DefFn(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "dms::Variable FN";
 	std::string option = FixOptionNumber(lexical.option);
@@ -1893,7 +1893,7 @@ std::string MZ1Z001::DefFn(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-void MZ1Z001::EndDefFn(int number, int& subNumber)
+void MZ1Z002::EndDefFn(int number, int& subNumber)
 {
 	this->result += "; };";
 	this->grobal.push_back(this->result);
@@ -1901,7 +1901,7 @@ void MZ1Z001::EndDefFn(int number, int& subNumber)
 	this->defFnFlag = false;
 }
 
-void MZ1Z001::Input(const Lexical& lexical, bool delimiter, int number, int& subNumber)
+void MZ1Z002::Input(const Lexical& lexical, bool delimiter, int number, int& subNumber)
 {
 	std::string option = Trim(lexical.option);
 	bool showQuestion = true;
@@ -1941,7 +1941,7 @@ void MZ1Z001::Input(const Lexical& lexical, bool delimiter, int number, int& sub
 	this->result = "";
 }
 
-std::string MZ1Z001::Restore(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Restore(const Lexical& lexical, bool delimiter)
 {
 	std::string result;
 	if(lexical.option.empty() == true)
@@ -1960,13 +1960,13 @@ std::string MZ1Z001::Restore(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Clr(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Clr(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Clr()";
 	return result;
 }
 
-std::string MZ1Z001::Music(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Music(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixPrintOption(lexical.option, delimiter, true);
@@ -1974,7 +1974,7 @@ std::string MZ1Z001::Music(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Tempo(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Tempo(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -1982,7 +1982,7 @@ std::string MZ1Z001::Tempo(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-void MZ1Z001::CloseBracket(int number, int& subNumber)
+void MZ1Z002::CloseBracket(int number, int& subNumber)
 {
 	this->result += ");";
 	this->program[Number(number, subNumber)] = this->result;
@@ -1991,7 +1991,7 @@ void MZ1Z001::CloseBracket(int number, int& subNumber)
 	this->closeBracketFlag = false;
 }
 
-std::string MZ1Z001::Usr(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Usr(const Lexical& lexical, bool delimiter)
 {
 //	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -1999,7 +1999,7 @@ std::string MZ1Z001::Usr(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Wopen(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Wopen(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2007,7 +2007,7 @@ std::string MZ1Z001::Wopen(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Ropen(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Ropen(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2015,19 +2015,19 @@ std::string MZ1Z001::Ropen(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Close(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Close(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Close()";
 	return result;
 }
 
-std::string MZ1Z001::Mon(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Mon(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Mon()";
 	return result;
 }
 
-std::string MZ1Z001::Limit(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Limit(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2040,13 +2040,13 @@ std::string MZ1Z001::Limit(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Cont(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Cont(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Cont()";
 	return result;
 }
 
-std::string MZ1Z001::Get(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Get(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = Trim(lexical.option);
@@ -2054,7 +2054,7 @@ std::string MZ1Z001::Get(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Inp(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Inp(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2062,7 +2062,7 @@ std::string MZ1Z001::Inp(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Out(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Out(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2070,7 +2070,7 @@ std::string MZ1Z001::Out(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Cursor(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Cursor(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2078,7 +2078,7 @@ std::string MZ1Z001::Cursor(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Set(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Set(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	this->addColorFlag = true;
@@ -2087,7 +2087,7 @@ std::string MZ1Z001::Set(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Reset(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Reset(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	this->addColorFlag = true;
@@ -2096,7 +2096,7 @@ std::string MZ1Z001::Reset(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Line(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Line(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	this->closeBracesFlag = true;
@@ -2106,7 +2106,7 @@ std::string MZ1Z001::Line(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Bline(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Bline(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	this->closeBracesFlag = true;
@@ -2116,7 +2116,7 @@ std::string MZ1Z001::Bline(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-void MZ1Z001::Console(const Lexical& lexical, bool delimiter, int number, int& subNumber)
+void MZ1Z002::Console(const Lexical& lexical, bool delimiter, int number, int& subNumber)
 {
 	std::string option = Trim(lexical.option);
 	int phase = 0;
@@ -2281,7 +2281,7 @@ void MZ1Z001::Console(const Lexical& lexical, bool delimiter, int number, int& s
 	this->result = "";
 }
 
-void MZ1Z001::Graph(const Lexical& lexical, bool delimiter, int number, int& subNumber)
+void MZ1Z002::Graph(const Lexical& lexical, bool delimiter, int number, int& subNumber)
 {
 	std::string command;
 	std::string page;
@@ -2371,7 +2371,7 @@ void MZ1Z001::Graph(const Lexical& lexical, bool delimiter, int number, int& sub
 	this->result = "";
 }
 
-std::string MZ1Z001::Position(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Position(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2379,7 +2379,7 @@ std::string MZ1Z001::Position(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Pattern(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Pattern(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	this->patternFlag = false;
@@ -2390,7 +2390,7 @@ std::string MZ1Z001::Pattern(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Auto(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Auto(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2398,7 +2398,7 @@ std::string MZ1Z001::Auto(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Image(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Image(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2406,7 +2406,7 @@ std::string MZ1Z001::Image(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Copy(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Copy(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2414,7 +2414,7 @@ std::string MZ1Z001::Copy(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Page(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Page(const Lexical& lexical, bool delimiter)
 {
 	this->closeBracketFlag = true;
 	std::string option = FixOptionNumber(lexical.option);
@@ -2422,37 +2422,37 @@ std::string MZ1Z001::Page(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Boot(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Boot(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Boot()";
 	return result;
 }
 
-std::string MZ1Z001::Klist(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Klist(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Klist()";
 	return result;
 }
 
-std::string MZ1Z001::Change(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Change(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Change()";
 	return result;
 }
 
-std::string MZ1Z001::Rew(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Rew(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Rew()";
 	return result;
 }
 
-std::string MZ1Z001::Fast(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Fast(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Fast()";
 	return result;
 }
 
-std::string MZ1Z001::BasicFunction(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::BasicFunction(const Lexical& lexical, bool delimiter)
 {
 	std::string functionName = lexical.command;
 	std::transform(functionName.begin(), functionName.end(), functionName.begin(), tolower);
@@ -2461,7 +2461,7 @@ std::string MZ1Z001::BasicFunction(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::BasicStringFunction(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::BasicStringFunction(const Lexical& lexical, bool delimiter)
 {
 	std::string functionName = lexical.command;
 	std::transform(functionName.begin(), functionName.end(), functionName.begin(), tolower);
@@ -2477,7 +2477,7 @@ std::string MZ1Z001::BasicStringFunction(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::BasicSign(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::BasicSign(const Lexical& lexical, bool delimiter)
 {
 	std::string command = lexical.command;
 	if(command == "<>")
@@ -2500,13 +2500,13 @@ std::string MZ1Z001::BasicSign(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Pi(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Pi(const Lexical& lexical, bool delimiter)
 {
 	std::string result = "Pi()" + FixOptionNumber(lexical.option);
 	return result;
 }
 
-std::string MZ1Z001::Pos(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Pos(const Lexical& lexical, bool delimiter)
 {
 	std::string command = lexical.option.substr(0, 1).c_str();
 	std::transform(command.begin(), command.end(), command.begin(), tolower);
@@ -2515,7 +2515,7 @@ std::string MZ1Z001::Pos(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Color(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Color(const Lexical& lexical, bool delimiter)
 {
 	this->wMode = 0;
 	this->closeBracketFlag = true;
@@ -2525,7 +2525,7 @@ std::string MZ1Z001::Color(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::CColor(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::CColor(const Lexical& lexical, bool delimiter)
 {
 	this->wMode = 0;
 	this->closeBracketFlag = true;
@@ -2534,7 +2534,7 @@ std::string MZ1Z001::CColor(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Circle(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Circle(const Lexical& lexical, bool delimiter)
 {
 	this->wMode = 0;
 	this->closeBracketFlag = true;
@@ -2545,7 +2545,7 @@ std::string MZ1Z001::Circle(const Lexical& lexical, bool delimiter)
 }
 
 
-std::string MZ1Z001::Box(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Box(const Lexical& lexical, bool delimiter)
 {
 	this->wMode = 0;
 	this->closeBracketFlag = true;
@@ -2578,7 +2578,7 @@ std::string MZ1Z001::Box(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::string MZ1Z001::Paint(const Lexical& lexical, bool delimiter)
+std::string MZ1Z002::Paint(const Lexical& lexical, bool delimiter)
 {
 	this->wMode = 0;
 	this->closeBracketFlag = true;
@@ -2607,7 +2607,7 @@ std::string MZ1Z001::Paint(const Lexical& lexical, bool delimiter)
 	return result;
 }
 
-std::vector<std::string> MZ1Z001::SpritText(std::string text, std::string separator)
+std::vector<std::string> MZ1Z002::SpritText(std::string text, std::string separator)
 {
 	int bracketsCount = 0;
 	bool quotation = false;
@@ -2656,7 +2656,7 @@ std::vector<std::string> MZ1Z001::SpritText(std::string text, std::string separa
 	return spritTextList;
 }
 
-std::string MZ1Z001::FixCcolorOption(std::string option)
+std::string MZ1Z002::FixCcolorOption(std::string option)
 {
 	std::string fixOption = "";
 	std::vector<std::string> spritOptionList = SpritText(option, ",");
@@ -2696,7 +2696,7 @@ std::string MZ1Z001::FixCcolorOption(std::string option)
 }
 
 // colorCommand true: Color文解析, false: []に囲まれた色指定
-std::string MZ1Z001::FixColorOption(std::string option, bool colorCommand)
+std::string MZ1Z002::FixColorOption(std::string option, bool colorCommand)
 {
 	int bracketsCount = 0;
 	bool quotation = false;
@@ -2787,7 +2787,7 @@ std::string MZ1Z001::FixColorOption(std::string option, bool colorCommand)
 	return fixOption;
 }
 
-bool MZ1Z001::CheckEncode(unsigned char byte)
+bool MZ1Z002::CheckEncode(unsigned char byte)
 {
 	if(byte < 0x20)
 	{
@@ -2812,7 +2812,7 @@ bool MZ1Z001::CheckEncode(unsigned char byte)
 	return false;
 }
 
-bool MZ1Z001::CheckEncodeAfter(unsigned char byte)
+bool MZ1Z002::CheckEncodeAfter(unsigned char byte)
 {
 	if(byte >= 'A' && byte <= 'F')
 	{
@@ -2829,7 +2829,7 @@ bool MZ1Z001::CheckEncodeAfter(unsigned char byte)
 	return false;
 }
 
-std::string MZ1Z001::Encode(std::string text)
+std::string MZ1Z002::Encode(std::string text)
 {
 	std::string option = text;
 	std::string result;
@@ -2857,7 +2857,7 @@ std::string MZ1Z001::Encode(std::string text)
 	return result;
 }
 
-std::string MZ1Z001::Space(void)
+std::string MZ1Z002::Space(void)
 {
 	if(this->space == 0)
 	{
@@ -2866,18 +2866,18 @@ std::string MZ1Z001::Space(void)
 	return std::string(this->space, ' ');
 }
 
-std::string MZ1Z001::Number(int number, int subNumber)
+std::string MZ1Z002::Number(int number, int subNumber)
 {
 	return Format("%05d_%02d", number, subNumber);
 }
 
-std::string MZ1Z001::Number(std::string number)
+std::string MZ1Z002::Number(std::string number)
 {
 	int intNumber = atoi(number.c_str());
 	return Number(intNumber, 1);
 }
 
-std::string MZ1Z001::NumberList(std::string number)
+std::string MZ1Z002::NumberList(std::string number)
 {
 	std::string separator = ",";
 	size_t separatorLength = separator.length();
@@ -2908,7 +2908,7 @@ std::string MZ1Z001::NumberList(std::string number)
 	return result;
 }
 
-bool MZ1Z001::CheckReserved(std::string text)
+bool MZ1Z002::CheckReserved(std::string text)
 {
 	std::vector<std::string> disableList = {"TI$", "CSRV", "CSRH", "POSV", "POSH", "SIZE"};
 	for(size_t i = 0; i < disableList.size(); ++ i)
@@ -2921,7 +2921,7 @@ bool MZ1Z001::CheckReserved(std::string text)
 	return false;
 }
 
-std::string MZ1Z001::ReplaceAll(const std::string& text, const std::string& target, const std::string& replacement)
+std::string MZ1Z002::ReplaceAll(const std::string& text, const std::string& target, const std::string& replacement)
 {
 	std::string result = text;
 	std::string::size_type pos = 0;
@@ -2933,7 +2933,7 @@ std::string MZ1Z001::ReplaceAll(const std::string& text, const std::string& targ
 	return result;
 }
 
-bool MZ1Z001::CheckBasicFunction(const std::string& text)
+bool MZ1Z002::CheckBasicFunction(const std::string& text)
 {
 	std::vector<std::string> functionList = {"RND(", "SIN(", "COS(", "TAN(", "ATN(", "EXP(", "INT(", "LOG(", "LN(", "ABS(", "SGN(", "SQR(", "ASC(", "VAL(", "POINT(", "LEN(", "PEEK(", "TAB("};
 	for(size_t i = 0; i < functionList.size(); ++ i)
@@ -2946,7 +2946,7 @@ bool MZ1Z001::CheckBasicFunction(const std::string& text)
 	return false;
 }
 
-bool MZ1Z001::CheckBasicStringFunction(const std::string& text)
+bool MZ1Z002::CheckBasicStringFunction(const std::string& text)
 {
 	std::vector<std::string> functionList = {"LEFT$(", "RIGHT$(", "MID$(", "CHR$(", "STR$(", "SPACE$(", "STRING$(", "CHARACTER$("};
 	for (size_t i = 0; i < functionList.size(); ++i)
@@ -2960,7 +2960,7 @@ bool MZ1Z001::CheckBasicStringFunction(const std::string& text)
 }
 
 // 不等号かどうか
-bool MZ1Z001::CheckBasicSign(const std::string& text)
+bool MZ1Z002::CheckBasicSign(const std::string& text)
 {
 	std::vector<std::string> functionList = { "><", "<>", "=<", "<=", "=>", ">=", ">", "<"};
 	for (size_t i = 0; i < functionList.size(); ++i)
@@ -2974,7 +2974,7 @@ bool MZ1Z001::CheckBasicSign(const std::string& text)
 }
 
 // データを文字列のarrayに変換する
-std::string MZ1Z001::ParseData(const std::string& data)
+std::string MZ1Z002::ParseData(const std::string& data)
 {
 	std::string sourceData = data;
 	if(data.empty() == true)
@@ -3053,7 +3053,7 @@ std::string MZ1Z001::ParseData(const std::string& data)
 
 // 文字列変数名を変更する
 // A$ => strA
-std::string MZ1Z001::ChangeVariableArrayName(std::string variableName)
+std::string MZ1Z002::ChangeVariableArrayName(std::string variableName)
 {
 	variableName = "array" + variableName;
 	return variableName;
@@ -3061,7 +3061,7 @@ std::string MZ1Z001::ChangeVariableArrayName(std::string variableName)
 
 // 文字列変数名を変更する
 // A$ => strA
-std::string MZ1Z001::ChangeStringVariableName(std::string variableName, bool array)
+std::string MZ1Z002::ChangeStringVariableName(std::string variableName, bool array)
 {
 	std::string prefix = "";
 	size_t dollarPos = variableName.find_last_of('$');
@@ -3087,7 +3087,7 @@ std::string MZ1Z001::ChangeStringVariableName(std::string variableName, bool arr
 }
 
 // 文字列変数名($付き)かどうか
-bool MZ1Z001::IsStringVariableName(std::string variableName)
+bool MZ1Z002::IsStringVariableName(std::string variableName)
 {
 	size_t dollarPos = variableName.find_last_of('$');
 	if(dollarPos == 0)
@@ -3102,7 +3102,7 @@ bool MZ1Z001::IsStringVariableName(std::string variableName)
 }
 
 // 文字列変数名(str付き)かどうか
-bool MZ1Z001::IsStringFixedVariableName(std::string variableName)
+bool MZ1Z002::IsStringFixedVariableName(std::string variableName)
 {
 	size_t strPos = variableName.find("str");
 	if(strPos != std::string::npos)
@@ -3113,7 +3113,7 @@ bool MZ1Z001::IsStringFixedVariableName(std::string variableName)
 }
 
 // 文字列からスペースを削除する
-std::string MZ1Z001::EraseSpace(const std::string& text)
+std::string MZ1Z002::EraseSpace(const std::string& text)
 {
 	std::string result = "";
 	bool quotation = false;
@@ -3140,7 +3140,7 @@ std::string MZ1Z001::EraseSpace(const std::string& text)
 }
 
 // 前後のスペースを削除する
-std::string MZ1Z001::Trim(const std::string& text, const char* trimCharacterList)
+std::string MZ1Z002::Trim(const std::string& text, const char* trimCharacterList)
 {
 	std::string result;
 	std::string::size_type left = text.find_first_not_of(trimCharacterList);
@@ -3153,7 +3153,7 @@ std::string MZ1Z001::Trim(const std::string& text, const char* trimCharacterList
 }
 
 // ,で区切られた変数リストを取得する
-std::vector<std::string> MZ1Z001::VariableList(const std::string& text)
+std::vector<std::string> MZ1Z002::VariableList(const std::string& text)
 {
 	std::string separator = ",";
 	size_t separatorLength = separator.length();
@@ -3180,7 +3180,7 @@ std::vector<std::string> MZ1Z001::VariableList(const std::string& text)
 }
 
 // Dim関数用,で区切られた変数リストを取得する
-std::vector<std::string> MZ1Z001::DimVariableList(const std::string& text)
+std::vector<std::string> MZ1Z002::DimVariableList(const std::string& text)
 {
 	std::vector<std::string> variableList = VariableList(text);
 	// 2次元配列を修復する
@@ -3204,7 +3204,7 @@ std::vector<std::string> MZ1Z001::DimVariableList(const std::string& text)
 }
 
 // ダブルクォーテーションで囲まれた文字列をスペースで埋める
-std::string MZ1Z001::DeleteQuotationString(std::string& text)
+std::string MZ1Z002::DeleteQuotationString(std::string& text)
 {
 	std::string result = "";
 	bool quotation = false;
@@ -3226,7 +3226,7 @@ std::string MZ1Z001::DeleteQuotationString(std::string& text)
 	return result;
 }
 
-void MZ1Z001::ReplaceNumber(void)
+void MZ1Z002::ReplaceNumber(void)
 {
 	// 次の命令の行番号と次の行番号を抽出
 	auto dataBegin = this->data.begin();
@@ -3271,7 +3271,7 @@ void MZ1Z001::ReplaceNumber(void)
 	}
 }
 
-std::string MZ1Z001::ReplaceAll(std::string &replacedStr, std::string from, std::string to)
+std::string MZ1Z002::ReplaceAll(std::string &replacedStr, std::string from, std::string to)
 {
 	size_t pos = replacedStr.find(from);
 	size_t toLen = to.length();
@@ -3288,7 +3288,7 @@ std::string MZ1Z001::ReplaceAll(std::string &replacedStr, std::string from, std:
 }
 
 // 変数リストに含まれるかどうか
-bool MZ1Z001::FindVariableList(std::string variableName)
+bool MZ1Z002::FindVariableList(std::string variableName)
 {
 	for(size_t i = 0; i < this->variableList.size(); ++ i)
 	{
@@ -3307,7 +3307,7 @@ bool MZ1Z001::FindVariableList(std::string variableName)
 	return false;
 }
 
-bool MZ1Z001::FindDefFnFuncList(std::string variableName)
+bool MZ1Z002::FindDefFnFuncList(std::string variableName)
 {
 	for(size_t i = 0; i < this->deffnFuncList.size(); ++ i)
 	{
@@ -3320,14 +3320,14 @@ bool MZ1Z001::FindDefFnFuncList(std::string variableName)
 }
 
 // オプションを修正する
-std::string MZ1Z001::FixOption(std::string sourceCommand, std::string sourceOption, bool delimiter)
+std::string MZ1Z002::FixOption(std::string sourceCommand, std::string sourceOption, bool delimiter)
 {
 	std::string option = Trim(sourceOption);
 	return option;
 }
 
 // Print文のセミコロンを+に変換する
-std::string MZ1Z001::FixPrintOption(std::string sourceOption, bool delimiter, bool commaToPlus)
+std::string MZ1Z002::FixPrintOption(std::string sourceOption, bool delimiter, bool commaToPlus)
 {
 	std::string result;
 	std::string option = FixOptionNumber(sourceOption);
@@ -3375,7 +3375,7 @@ std::string MZ1Z001::FixPrintOption(std::string sourceOption, bool delimiter, bo
 	return result;
 }
 
-std::string MZ1Z001::FixCircleOption(std::string sourceOption)
+std::string MZ1Z002::FixCircleOption(std::string sourceOption)
 {
 	std::string result;
 	std::string option = FixOptionNumber(sourceOption);
@@ -3404,7 +3404,7 @@ std::string MZ1Z001::FixCircleOption(std::string sourceOption)
 
 // 数値リテラルに_nまたは_fをつける
 // $####も16進数の数値リテラル扱いにして\xを先頭につける
-std::string MZ1Z001::FixOptionNumber(std::string sourceOption)
+std::string MZ1Z002::FixOptionNumber(std::string sourceOption)
 {
 	std::string result;
 	std::string option = Trim(sourceOption);
@@ -3510,7 +3510,7 @@ std::string MZ1Z001::FixOptionNumber(std::string sourceOption)
 	return result;
 }
 
-void MZ1Z001::DebugLog(std::string text)
+void MZ1Z002::DebugLog(std::string text)
 {
 	FileData file;
 	file.SetBuffer(&text[0], text.size());
